Policy checks in Network, User and Practice Controllers

+++ app/Http/Controllers/Practice/PracticeController.php
@@ -38,7 +38,10 @@ class PracticeController extends Controller
      */
     public function create()
     {
-        $this->authorize('add-practices');
+        if(!policy(new Practice)->administration()){
+            session()->flash('failure', 'Unauthorized Access!');
+            return redirect('/home');
+        }
 
         $id = -1;
         $data = array();
@@ -63,6 +66,11 @@ class PracticeController extends Controller
      */
     public function store(Request $request)
     {
+        if(!policy(new Practice)->administration()){
+            session()->flash('failure', 'Unauthorized Access!');
+            return redirect('/home');
+        }
+
         $practicedata = json_decode($request->input('data'), true);
         $practice = new Practice;
         $practice->name = $practicedata[0]['practice_name'];
@@ -115,7 +123,6 @@ class PracticeController extends Controller
      */
     public function show(Request $request)
     {
-        $this->authorize('add-practices');
         $data = array();
         $practice_id = $request->input('practice_id');
         $practice_name = Practice::find($practice_id)->name;
@@ -139,6 +146,11 @@ class PracticeController extends Controller
      */
     public function edit($id, $location)
     {
+        if(!policy(new Practice)->administration()){
+            session()->flash('failure', 'Unauthorized Access!');
+            return redirect('/home');
+        }
+
         $data = array();
         $data['practice_active'] = true;
         $data['id'] = $id;
@@ -161,6 +173,11 @@ class PracticeController extends Controller
      */
     public function update(Request $request)
     {
+        if(!policy(new Practice)->administration()){
+            session()->flash('failure', 'Unauthorized Access!');
+            return redirect('/home');
+        }
+
         $practicedata = json_decode($request->input('data'), true);
         $practicename = $practicedata[0]['practice_name'];
         $practiceemail = $practicedata[0]['practice_email'];
@@ -220,6 +237,11 @@ class PracticeController extends Controller
      */
     public function destroy(Request $request)
     {
+        if(!policy(new Practice)->administration()){
+            session()->flash('failure', 'Unauthorized Access!');
+            return redirect('/home');
+        }
+
         $i = 0;
         while (1) {
             if ($request->input($i)) {
@@ -267,7 +289,11 @@ class PracticeController extends Controller
 
     public function administration(Request $request)
     {
-        $this->authorize('add-practices');
+
+        if(!policy(new Practice)->administration()){
+            session()->flash('failure', 'Unauthorized Access!');
+            return redirect('/home');
+        }
         $data = array();
         $data['practice_active'] = true;
         return view('practice.admin')->with('data', $data);
@@ -275,6 +301,11 @@ class PracticeController extends Controller
 
     public function removelocation(Request $request)
     {
+        if(!policy(new Practice)->administration()){
+            session()->flash('failure', 'Unauthorized Access!');
+            return redirect('/home');
+        }
+        
         $data = PracticeLocation::find($request->location_id)->delete();
 
         return json_encode($data);
